fix aws-chunked usage of JVM platform code in common sourceSets
