--- runtime/auth/aws-signing-tests/common/src/aws/smithy/kotlin/runtime/auth/awssigning/tests/AwsChunkedSourceTestBase.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package aws.smithy.kotlin.runtime.auth.awssigning.tests
-
-public abstract class AwsChunkedSourceTestBase : AwsChunkedTestBase(AwsChunkedReaderFactory.Source)
