+++ runtime/io/common/src/aws/smithy/kotlin/runtime/io/SdkByteReadChannel.kt
@@ -4,6 +4,9 @@
  */
 package aws.smithy.kotlin.runtime.io
 
+import aws.smithy.kotlin.runtime.io.internal.SdkDispatchers
+import kotlinx.coroutines.withContext
+
 /**
  * Supplies an asynchronous stream of bytes. This is a **single-reader channel**.
  */
@@ -96,4 +99,35 @@ public suspend fun SdkByteReadChannel.readToBuffer(): SdkBuffer {
     val buffer = SdkBuffer()
     readRemaining(buffer)
     return buffer
+}
+
+/**
+ * Read all bytes from this channel into [sink]. Returns the total number of bytes written.
+ */
+public suspend fun SdkByteReadChannel.readAll(sink: SdkSink): Long = withContext(SdkDispatchers.IO) {
+    val bufferedSink = if (sink is SdkBuffer) sink else sink.buffer()
+    var totalWritten = 0L
+    while (true) {
+        val rc = read(bufferedSink.buffer, DEFAULT_BYTE_CHANNEL_MAX_BUFFER_SIZE.toLong())
+        if (rc == -1L) break
+        totalWritten += rc
+        bufferedSink.emit()
+    }
+    bufferedSink.emit()
+    totalWritten
+}
+
+/**
+ * Removes all bytes from [source] and writes them to this channel. Returns the total number of bytes read.
+ */
+public suspend fun SdkByteWriteChannel.writeAll(source: SdkSource): Long = withContext(SdkDispatchers.IO) {
+    val buffer = SdkBuffer()
+    var totalRead = 0L
+    while (true) {
+        val rc = source.read(buffer, DEFAULT_BYTE_CHANNEL_MAX_BUFFER_SIZE.toLong())
+        if (rc == -1L) break
+        totalRead += rc
+        write(buffer)
+    }
+    totalRead
 }
